% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookend_line.R
\name{bookend_line}
\alias{bookend_line}
\title{Internal function for formatting a line with characters on left and right sides.}
\usage{
bookend_line(x, linewidth = 80, type = c("basic", "roxygen2"))
}
\arguments{
\item{x}{character string.}

\item{linewidth}{Intended maximum number of characters per line.}

\item{type}{Two options:
\itemize{
\item{\code{basic}: Simple format with \code{#} box at the beginning.}
\item{\code{roxygen2}: Format consistent with package builder \code{roxygen2}.}}}
}
\value{
Character string of length 80.
}
\description{
Internal function for formatting a line with characters on left and right sides.
}
\examples{
bookend_line("abcde",linewidth = 10,type = "basic")
bookend_line("abcde",linewidth = 10,type = "roxygen2")
}
