% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_run.R
\name{check_files_run}
\alias{check_files_run}
\title{Check to see if the file will run to completion without error:}
\usage{
check_files_run(files, verbose = FALSE, line_selection = NULL)
}
\arguments{
\item{files}{character vector of R-script files to be checked by running in order.}

\item{verbose}{Print messages for checking progress?}

\item{line_selection}{Select the lines to be run if only a subset to be used. Default is for all lines to be run, and user should be very careful with this option, as subsequent code execution may be affected by changes to prior code.}
}
\description{
Function takes an R-script file as input and attempts to run the script in a new console R-session.
}
