% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_script_title_box.R
\name{create_r_script_title_box}
\alias{create_r_script_title_box}
\title{Create an R script file with preamble in either the basic or roxygen2 format.}
\usage{
create_r_script_title_box(
  title,
  author,
  email,
  date_created = Sys.Date(),
  date_last_modified = Sys.Date(),
  description,
  params = NULL,
  output = NULL,
  license = "MIT",
  linewidth = 80,
  right_margin = 4,
  left_margin = 1,
  type = c("basic", "roxygen2"),
  write_to_file = NULL
)
}
\arguments{
\item{title}{Character string with title of script.}

\item{author}{Author's name}

\item{email}{Author's email address}

\item{date_created}{Date the script file was created. Defaults to today.}

\item{date_last_modified}{Date the script file was last modified. Defaults to today.}

\item{description}{Character string description of the script's function.}

\item{params}{A vector of strings detailing the parameters (most useful if script is a function.)}

\item{output}{A string detailing the output of the script.}

\item{license}{A string detailing the license for the code.}

\item{linewidth}{Maximum number of characters per line.}

\item{right_margin}{Number of spaces between the end of text and the maximum linewidth.}

\item{left_margin}{Number of spaces before the beginning of text.}

\item{type}{Two options:
\itemize{
\item{\code{basic}: Simple format with \code{#} box at the beginning.}
\item{\code{roxygen2}: Format consistent with package builder \code{roxygen2}.}}}

\item{write_to_file}{File to create with preamble. If \code{NULL}, then outputted to the console.}
}
\value{
If file runs cleanly, then output is \code{TRUE}.
}
\description{
Create an R script file with preamble in either the basic or roxygen2 format.
}
\examples{
create_r_script_title_box(
  title = "A script about nothing",
  author = "D. W. Kennedy",
  email = "not.my.email@gmail.com",
  license = "",
  description = "\nThe following code:\n - solves existential dread,\n - the problem of evil\n - the need for accountants.",
  type = "basic",
  params = c("blah a parameter which is not important.", "verbose Should progress messages be shown?")
)
}
