% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Rdata.R
\name{load_Rdata}
\alias{load_Rdata}
\title{Function to safely load Rdata files.}
\usage{
load_Rdata(file)
}
\arguments{
\item{file}{string file path to .Rdata object.}
}
\value{
a list containing all objects contained in Rdata file.
}
\description{
The \code{load()} function loads the objects stored in the Rdata file directly into the environment, which can write over existing objects. This function allows the objects to be assigned to a list variable. Since Rdata files are now largely replaced by the safer RDS format, this is essentially a back-compatibility mechanism.
}
