% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_white_space.R
\name{strip_white_space}
\alias{strip_white_space}
\title{Remove leading and trailing white spaces.}
\usage{
strip_white_space(x)
}
\arguments{
\item{x}{String.}
}
\value{
string with no leading or trailing white space.
}
\description{
Remove leading and trailing white spaces.
}
\examples{
strip_white_space("  sdfs adf ")
}
