% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncomment_marked_lines.R
\name{uncomment_marked_lines}
\alias{uncomment_marked_lines}
\title{Function to uncomment all commented-out lines.}
\usage{
uncomment_marked_lines(
  file,
  new_file_name,
  uncomment_pattern = "##UNCOMMENT##$"
)
}
\arguments{
\item{file}{character string with file path to .R file.}

\item{new_file_name}{character string with file path (and name) for the output .R file.}

\item{uncomment_pattern}{Regular expression for finding uncomment marker. Any lines containing this pattern are uncommented.}
}
\value{
If file runs cleanly, then output is \code{TRUE}.
}
\description{
Sometimes commands which work on the client side do not work or are detrimental on your side. (Think install.packages lines, or file system commands.)
This function takes an .R script file and uncomments any lines previously marked for uncommenting. Lines can be added to your version with a pattern e.g. \code{##UNCOMMENT##}, and then removed when the code is shipped back to the client.
}
