% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_r_script_title_box.R
\name{update_r_script_title_box}
\alias{update_r_script_title_box}
\title{Update the title box's "date last modified" field.}
\usage{
update_r_script_title_box(file, new_date = Sys.Date())
}
\arguments{
\item{file}{Path to .R script file to be updated.}

\item{new_date}{The date to be given as the last modification date. Default is today.}
}
\value{
If file runs cleanly, then output is \code{TRUE}.
}
\description{
Useful function to automatically update the last modified field when changes are made to a script.
This is a simple implementation of version control, and a modern system such as 'git' would be recommended instead.
}
